--- base/src/components/kystverket/ApplicationHeader/ApplicationHeader.module.css
@@ -1,76 +1,0 @@
-@layer styrbord-modules {
-  .logo {
-    flex-shrink: 0;
-    @container header (width < 37.5rem) {
-      flex-shrink: 1;
-    }
-  }
-
-  .logoLink {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 14px;
-    line-height: 0;
-    text-decoration: none !important;
-
-    &:active {
-      background-color: inherit;
-    }
-  }
-
-  .titleText {
-    display: inline-block;
-    color: var(--ds-color-primary-base-default);
-    font-weight: lighter;
-    font-size: var(--ds-font-size-4);
-    letter-spacing: 0.1em !important;
-    text-transform: uppercase;
-  }
-  .headerContainer {
-    container: header / inline-size;
-    position: relative;
-    height: fit-content;
-    min-height: 72px;
-    border-bottom: 1px solid var(--ds-color-neutral-surface-tinted);
-    border-bottom-right-radius: var(--ds-border-radius-small);
-    border-bottom-left-radius: var(--ds-border-radius-small);
-    background-color: var(--ds-color-neutral-background-default);
-  }
-  .mainLinkButton {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: flex-start;
-    padding: var(--ds-size-3) var(--ds-size-4);
-    gap: var(--ds-size-1);
-    border-radius: var(--ds-border-radius-md);
-    color: var(--ds-color-text-default);
-    font-weight: var(--ds-font-weight-medium);
-    text-decoration: none;
-    text-transform: none;
-    white-space: nowrap;
-    &:hover {
-      background-color: var(--ds-color-surface-hover);
-    }
-    &:active {
-      background-color: var(--ds-color-surface-active);
-    }
-  }
-
-  @container header (width < 64rem) {
-    .smallScreenHide {
-      display: none;
-    }
-  }
-
-  @container header (width >= 64rem) {
-    .notSmallScreenHide {
-      display: none;
-    }
-  }
-
-  .dropdownButton {
-    padding: var(--ds-size-2) var(--ds-size-2);
-  }
-}
